+++ src/main/java/com/hubspot/jackson/datatype/protobuf/builtin/serializers/WrappedPrimitiveSerializer.java
@@ -1,7 +1,6 @@
 package com.hubspot.jackson.datatype.protobuf.builtin.serializers;
 
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParser.NumberType;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.SerializerProvider;
@@ -10,10 +9,11 @@ import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.MessageOrBuilder;
 import com.hubspot.jackson.datatype.protobuf.ProtobufJacksonConfig;
 import com.hubspot.jackson.datatype.protobuf.ProtobufSerializer;
+import com.hubspot.jackson.datatype.protobuf.internal.FieldSchemaGenerator;
 import java.io.IOException;
-import java.math.BigInteger;
 
 public class WrappedPrimitiveSerializer<T extends MessageOrBuilder> extends ProtobufSerializer<T> {
+  private final FieldDescriptor valueField;
 
   /**
    * @deprecated use {@link #WrappedPrimitiveSerializer(Class, ProtobufJacksonConfig)}
@@ -23,8 +23,18 @@ public class WrappedPrimitiveSerializer<T extends MessageOrBuilder> extends Prot
     this(wrapperType, ProtobufJacksonConfig.getDefaultInstance());
   }
 
+  /**
+   * @deprecated use {@link #WrappedPrimitiveSerializer(T, ProtobufJacksonConfig)}
+   */
+  @Deprecated
   public WrappedPrimitiveSerializer(Class<T> wrapperType, ProtobufJacksonConfig config) {
-    super(wrapperType, config);
+    this(defaultInstance(wrapperType), config);
+  }
+
+  @SuppressWarnings("unchecked")
+  public WrappedPrimitiveSerializer(T defaultInstance, ProtobufJacksonConfig config) {
+    super((Class<T>) defaultInstance.getClass(), config);
+    this.valueField = defaultInstance.getDescriptorForType().findFieldByName("value");
   }
 
   @Override
@@ -33,55 +43,21 @@ public class WrappedPrimitiveSerializer<T extends MessageOrBuilder> extends Prot
           JsonGenerator generator,
           SerializerProvider serializerProvider
   ) throws IOException {
-    FieldDescriptor field = message.getDescriptorForType().findFieldByName("value");
-    Object value = message.getField(field);
-    writeValue(field, value, generator, serializerProvider);
+    Object value = message.getField(valueField);
+    writeValue(valueField, value, generator, serializerProvider);
   }
 
   @Override
   public void acceptJsonFormatVisitor(JsonFormatVisitorWrapper visitor, JavaType typeHint) throws JsonMappingException {
-    switch (handledType().getSimpleName()) {
-      case "StringValue":
-      case "BytesValue":
-        visitor.expectStringFormat(visitor.getProvider().constructType(String.class));
-        break;
-      case "BoolValue":
-        visitor.expectBooleanFormat(visitor.getProvider().constructType(Boolean.class));
-        break;
-      case "DoubleValue":
-        visitFloatFormat(visitor, visitor.getProvider().constructType(Double.class), NumberType.DOUBLE);
-        break;
-      case "FloatValue":
-        visitFloatFormat(visitor, visitor.getProvider().constructType(Float.class), NumberType.FLOAT);
-        break;
-      case "Int64Value":
-        if (getConfig().serializeLongsAsString()) {
-          visitor.expectStringFormat(visitor.getProvider().constructType(String.class));
-        } else {
-          visitIntFormat(visitor, visitor.getProvider().constructType(Long.class), NumberType.LONG);
-        }
-        break;
-      case "UInt64Value":
-        if (getConfig().serializeLongsAsString()) {
-          visitor.expectStringFormat(visitor.getProvider().constructType(String.class));
-        } else if (getConfig().properUnsignedNumberSerialization()) {
-          visitIntFormat(visitor, visitor.getProvider().constructType(BigInteger.class), NumberType.BIG_INTEGER);
-        } else {
-          visitIntFormat(visitor, visitor.getProvider().constructType(Long.class), NumberType.LONG);
-        }
-        break;
-      case "Int32Value":
-        visitIntFormat(visitor, visitor.getProvider().constructType(Integer.class), NumberType.INT);
-        break;
-      case "UInt32Value":
-        if (getConfig().properUnsignedNumberSerialization()) {
-          visitIntFormat(visitor, visitor.getProvider().constructType(Long.class), NumberType.LONG);
-        } else {
-          visitIntFormat(visitor, visitor.getProvider().constructType(Integer.class), NumberType.INT);
-        }
-        break;
-      default:
-        throw new IllegalStateException("Unexpected wrapper type: " + handledType());
+    new FieldSchemaGenerator(valueField, getConfig()).acceptJsonFormatVisitor(visitor, typeHint);
+  }
+
+  @SuppressWarnings("unchecked")
+  private static <T> T defaultInstance(Class<T> type) {
+    try {
+      return (T) type.getMethod("getDefaultInstance").invoke(null);
+    } catch (Exception e) {
+      throw new RuntimeException("Unable to get default instance for type " + type, e);
     }
   }
 }
